+++ src/basicop.cpp
@@ -20,7 +20,7 @@ std::vector<Real> FindPeaks(const std::vector<Real>& vector,
                             const Real min_peak_height) {
   std::vector<UInt> indexes = FindPeaksIndexes(vector, min_peak_height);
   std::vector<Real> output(indexes.size());
-  for (Int i=0; i<indexes.size(); ++i) { output[i] = vector[indexes[i]]; }
+  for (Int i=0; i<(Int)indexes.size(); ++i) { output[i] = vector[indexes[i]]; }
   return output;
 }
 
@@ -28,7 +28,7 @@ std::vector<UInt> FindPeaksIndexes(const std::vector<Real>& vector,
                                    const Real min_peak_height) {
   // Allocate new vectors for the indexes of the local maxima
   std::vector<UInt> indexes;
-  for (Int i=1; i<(vector.size()-1); ++i) {
+  for (Int i=1; i<(Int)(vector.size()-1); ++i) {
     if ((vector[i] > min_peak_height) &
         (vector[i] > vector[i-1]) &
         (vector[i] > vector[i+1])) {
